chore: extracting encode source map function from index [TOL-1788]
